pyproject.toml
[project]
name = "code"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.107.2",
    "pandas>=2.3.2",
    "pypdf>=6.0.0",
    "python-dotenv>=1.1.1",
    "torch>=2.8.0",
    "transformers>=4.56.1",
]
